requirements.txt
fastapi
uvicorn
pillow
requests
pytest
pyjwt
python-dotenv
python-multipart